MLytics_Studio/requirements_pip.txt
amqp==2.4.2
asgiref==1.1.2
attrs==19.1.0
autobahn==19.3.2
Automat==0.7.0
backports.csv==1.0.7
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-admin-tools==0.8.1
django-celery-results==1.0.4
django-channels==0.7.0
django-crispy-forms==1.7.0
django-extensions==2.1.6
django-import-export==1.2.0
django-jet==1.0.8
django-material==1.5.2
django-modelclone==0.7.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
kombu==4.4.0
numpy==1.16.2
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.1
pandas==0.24.2
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
python-dateutil==2.8.0
python-gitlab==1.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.1
tablib==0.13.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
